Move generation logic out of dispatch function
The `dispatch` function is intended to serve as a simple mapping from a
subcommand to the function that implements it. While it doesn't yet
contain any real logic, trying to do anything with the result of
`find_cue_files` will change that. Before that happens, though, I'm
adding the `generate_playlists` function to contain any and all logic
related to using the `generate` subcommand.