Strip source path from cue paths
When creating the `m3u` files, paths that are relative to it are needed
for the `cue` files. Right now we only have an absolute path. This will
strip out the source path, leaving us with a relative path instead.
These new paths can be dropped right into the `m3u` files.