Group cue files by folder
When generating `m3u` files recursively, it's important that not all
`cue` files get lumped together into a single playlist. This will add a
function, `group_files_by_folder`, that will get the prefix of each file
passed in and use that as a key by which to group the files.

There could be a problem with this approach if the `cue` files are
nested within multiple subfolders, but for right now I think this
approach will be sufficient.